+++ src/makeVariables.ts
@@ -1,6 +1,5 @@
 import { ColorValue } from './colors';
 import { RawCssString } from './css';
-import { resolveType } from './makeVariables.resolveType';
 import { UnitValue } from './units';
 
 type Mutable<T> = {
@@ -21,16 +20,15 @@ const DEFAULT_NAME_MAPPER = (name: string) => name;
 type VarsMap<T extends string> = Record<T, VariablePrinter>;
 
 export type VariableStyles<T extends string> = {
-  declarations: RawCssString;
   readonly vars: VarsMap<T>;
-  override<K extends T>(vars: Record<K, VariableValue>): RawCssString;
+  declare(vars: Record<T, VariableValue>): RawCssString;
+  override(vars: Partial<Record<T, VariableValue>>): RawCssString;
 };
 
 export type VariablePrinter = {
   (defaultValue?: VariableValue): RawCssVarString;
 } & Readonly<{
   cssName: string;
-  type: string;
   toString(): string;
   toJSON(): string;
   [IS_PRINTER]: true;
@@ -51,27 +49,11 @@ export type VariableOptions = {
    * Default: `(name) => name`
    */
   toCSSName: (name: string) => string;
-  /**
-   * Use this option if you're using a custom color manipulation library.
-   *
-   * Runs **after** the default color-detection logic, so it
-   * does not need to check for common CSS color string formats.
-   *
-   * Example: `(value) => value instance of MyColorClass`
-   */
-  isColor?: (value: unknown) => boolean;
-  /**
-   * Runs ahead of the default type-resolution to detect custom types.
-   *
-   * If a falsy type name is returned, the default type resolver runs
-   * as normal.
-   */
-  resolveType?: (value: unknown) => string | undefined;
 };
 
 // ---------------------------------------------------------------------------
 
-const makeVariablePrinter = (name: string, type: string) => {
+const makeVariablePrinter = (name: string) => {
   const cssName = `--${name}`;
   const varString = `var(${cssName})`;
 
@@ -81,34 +63,24 @@ const makeVariablePrinter = (name: string, type: string) => {
       : varString) as unknown as Mutable<VariablePrinter>;
   printer[IS_PRINTER] = true;
   printer.toString = printer.toJSON = () => varString;
-  printer.type = type;
   printer.cssName = cssName;
 
   return printer as VariablePrinter;
 };
 
-const mapObject = <T extends string, V>(
-  input: Record<T, unknown>,
-  makeValue: (name: string, value: unknown) => V
-): Record<T, V> =>
-  Object.fromEntries(
-    (Object.entries(input) as Array<[T, unknown]>).map(([name, value]) => [
-      name,
-      makeValue(name, value),
-    ])
-  ) as Record<T, V>;
-
 // ---------------------------------------------------------------------------
 
 const makeDeclarations = (
-  vars: Record<string, VariableValue>,
+  vars: Partial<Record<string, VariableValue>>,
   allowed: Record<string, VariablePrinter>
 ): string => {
   return Object.entries(vars)
     .map(([key, value]) => {
       const printer = allowed[key];
-      const valueStr = String(value).trim();
-      return printer ? `${printer.cssName}: ${valueStr};\n` : '';
+      if (!printer || value == null) {
+        return '';
+      }
+      return `${printer.cssName}: ${String(value).trim()};\n`;
     })
     .join('');
 };
@@ -131,30 +103,23 @@ const assertValidName = (name: string, nameRe = DEFAULT_NAME_RE) => {
 };
 
 export const makeVariables = <T extends string>(
-  input: Record<T, VariableValue>,
+  variableTokens: Array<T>,
   options: Partial<VariableOptions> = {}
 ): VariableStyles<T> => {
   assertValidNameRe(options.nameRe);
 
-  const {
-    nameRe = DEFAULT_NAME_RE,
-    toCSSName = DEFAULT_NAME_MAPPER,
-    resolveType: getCustomType,
-    isColor,
-  } = options;
+  const { nameRe, toCSSName = DEFAULT_NAME_MAPPER } = options;
 
-  const vars: VarsMap<T> = mapObject(input, (name, value) => {
-    assertValidName(name, nameRe);
-    let type = (getCustomType && getCustomType(value)) || resolveType(value);
-    if (isColor && type !== 'color' && isColor(value)) {
-      type = 'color';
-    }
-    return makeVariablePrinter(toCSSName(name), type);
-  });
+  const vars = Object.fromEntries(
+    variableTokens.map((name) => {
+      assertValidName(name, nameRe);
+      return [name, makeVariablePrinter(toCSSName(name))];
+    })
+  ) as VarsMap<T>;
 
   return {
-    declarations: makeDeclarations(input, vars),
     vars,
+    declare: (declarations) => makeDeclarations(declarations, vars),
     override: (overrides) => makeDeclarations(overrides, vars),
   };
 };
@@ -173,8 +138,8 @@ makeVariables.join = <VArr extends Array<VariableStyles<string>>>(
 > => {
   const vars = Object.assign({}, ...varDatas.map((d) => d.vars));
   return {
-    declarations: varDatas.map((d) => d.declarations).join(''),
     vars,
+    declare: (declarations) => makeDeclarations(declarations, vars),
     override: (overrides) => makeDeclarations(overrides, vars),
   };
 };